change from using which to using hash
which was noisy on a box without pv which lead me to want to change it.
which is also not recommended to be used in scripts since return code
and output isn't necessarily guaranteed. Since we're on bash its safer
to use hash, which is a built-in which will speed all this up.